pyproject.toml
[project]
name = "medical-wizard-mcp"
version = "0.1.0"
description = "Medical Wizard MCP server for clinical trial and publication intelligence"
requires-python = ">=3.11"
dependencies = [
    "httpx",
    "pydantic>=2.0",
    "python-dotenv",
    "fastmcp>=3.2.2",
    "starlette",
    "google-cloud-bigquery>=3.38.0",
]

[project.optional-dependencies]
dev = [
    "ruff",
    "pytest",
    "pytest-asyncio",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/Medical_Wizard_MCP"]

[tool.ruff]
line-length = 100
target-version = "py311"
